pyproject.toml
[project]
name = "consumer-insights"
version = "0.1.0"
description = "Consumer insights synthesizer that aggregates feedback across platforms and extracts themes, sentiment, and personas using AI."
requires-python = ">=3.11"
dependencies = [
    "praw>=7.7",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "httpx>=0.27",
    "google-api-python-client>=2.100",
    "anthropic>=0.40",
    "openai>=1.50",
    "pydantic>=2.5",
    "python-dotenv>=1.0",
    "python-pptx>=0.6.23",
    "openpyxl>=3.1",
    "rich>=13.7",
    "tenacity>=8.2",
    "click>=8.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
    "aiofiles>=23.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.5"]

[project.scripts]
consumer-insights = "main:main"
